Re-namespace dependency SG classes
