+++ osu.Framework.SourceGeneration.Tests/Verifiers/CSharpMultiPhaseSourceGeneratorVerifier_Test.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
+using osu.Framework.SourceGeneration.Generators.Dependencies;
 
 namespace osu.Framework.SourceGeneration.Tests.Verifiers
 {
