+++ osu.Framework.SourceGeneration/Generators/Dependencies/Data/BackgroundDependencyLoaderAttributeData.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
-namespace osu.Framework.SourceGeneration.Data
+namespace osu.Framework.SourceGeneration.Generators.Dependencies.Data
 {
     public readonly struct BackgroundDependencyLoaderAttributeData
     {
