+++ osu.Framework.SourceGeneration/Generators/Dependencies/Data/BackgroundDependencyLoaderParameterData.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.CodeAnalysis;
 
-namespace osu.Framework.SourceGeneration.Data
+namespace osu.Framework.SourceGeneration.Generators.Dependencies.Data
 {
     public readonly struct BackgroundDependencyLoaderParameterData
     {
