+++ osu.Framework.SourceGeneration/Generators/Dependencies/Data/CachedAttributeData.cs
@@ -5,7 +5,7 @@ using System;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
-namespace osu.Framework.SourceGeneration.Data
+namespace osu.Framework.SourceGeneration.Generators.Dependencies.Data
 {
     public readonly struct CachedAttributeData : IEquatable<CachedAttributeData>
     {
