+++ osu.Framework.SourceGeneration/Generators/Dependencies/Emitters/DependenciesFileEmitter.cs
@@ -9,7 +9,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace osu.Framework.SourceGeneration.Emitters
+namespace osu.Framework.SourceGeneration.Generators.Dependencies.Emitters
 {
     public class DependenciesFileEmitter
     {
