+++ osu.Framework.SourceGeneration/Generators/Dependencies/Emitters/ResolvedMemberEmitter.cs
@@ -4,19 +4,19 @@
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using osu.Framework.SourceGeneration.Data;
+using osu.Framework.SourceGeneration.Generators.Dependencies.Data;
 
-namespace osu.Framework.SourceGeneration.Emitters
+namespace osu.Framework.SourceGeneration.Generators.Dependencies.Emitters
 {
     /// <summary>
-    /// Emits the statement for a [Cached] attribute on a member of the class.
+    /// Emits the statement for a [Resolved] attribute.
     /// </summary>
-    public class CachedMemberEmitter : IStatementEmitter
+    public class ResolvedMemberEmitter : IStatementEmitter
     {
         private readonly DependenciesFileEmitter fileEmitter;
-        private readonly CachedAttributeData data;
+        private readonly ResolvedAttributeData data;
 
-        public CachedMemberEmitter(DependenciesFileEmitter fileEmitter, CachedAttributeData data)
+        public ResolvedMemberEmitter(DependenciesFileEmitter fileEmitter, ResolvedAttributeData data)
         {
             this.fileEmitter = fileEmitter;
             this.data = data;
@@ -25,13 +25,16 @@ namespace osu.Framework.SourceGeneration.Emitters
         public IEnumerable<StatementSyntax> Emit()
         {
             yield return SyntaxFactory.ExpressionStatement(
-                SyntaxHelpers.CacheDependencyInvocation(
-                    fileEmitter.Candidate.GlobalPrefixedTypeName,
+                SyntaxFactory.AssignmentExpression(
+                    SyntaxKind.SimpleAssignmentExpression,
                     createMemberAccessor(),
-                    data.GlobalPrefixedTypeName,
-                    data.Name,
-                    data.PropertyName
-                ));
+                    SyntaxHelpers.GetDependencyInvocation(
+                        fileEmitter.Candidate.GlobalPrefixedTypeName,
+                        data.GlobalPrefixedTypeName,
+                        data.CachedName,
+                        data.GlobalPrefixedParentTypeName,
+                        data.CanBeNull,
+                        true)));
         }
 
         private ExpressionSyntax createMemberAccessor()
@@ -42,7 +45,7 @@ namespace osu.Framework.SourceGeneration.Emitters
                     SyntaxFactory.CastExpression(
                         SyntaxFactory.ParseTypeName(fileEmitter.Candidate.GlobalPrefixedTypeName),
                         SyntaxFactory.IdentifierName(DependenciesFileEmitter.TARGET_PARAMETER_NAME))),
-                SyntaxFactory.IdentifierName(data.PropertyName!));
+                SyntaxFactory.IdentifierName(data.PropertyName));
         }
     }
 }
